refactor(icons): extract types
