+++ packages/icons/src/components/ScalarIconAirplane.vue
@@ -5,9 +5,8 @@ export default {
 }
 </script>
 <script lang="ts" setup>
-import { computed } from 'vue'
-
-import { useScalarIcon, type ScalarIconProps } from '../hooks/useScalarIcon'
+import { useScalarIcon } from '@/hooks'
+import type { ScalarIconProps } from '@/types'
 
 const props = defineProps<ScalarIconProps>()
