+++ packages/icons/vite.config.ts
@@ -1,10 +1,13 @@
-import { createViteBuildOptions } from '@scalar/build-tooling'
+import { alias, createViteBuildOptions } from '@scalar/build-tooling'
 import vue from '@vitejs/plugin-vue'
 import { defineConfig } from 'vite'
 import svgLoader from 'vite-svg-loader'
 
 // https://vitejs.dev/config/
 export default defineConfig({
+  resolve: {
+    alias: alias(import.meta.url),
+  },
   plugins: [
     vue(),
     // Ensure the viewBox is preserved
